% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace.R
\name{.correct_negative_control}
\alias{.correct_negative_control}
\title{compute negative control bias correction (intended for internal use)}
\usage{
.correct_negative_control(mu, data, input, control_label, seed = NULL)
}
\arguments{
\item{mu}{mu posterior samples}

\item{data}{data}

\item{input}{stan model input parameters}

\item{control_label}{control label}

\item{seed}{seed for negative control sampling}
}
\description{
compute negative control bias correction (intended for internal use)
}
