% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace.R
\name{.filter_data_low_counts}
\alias{.filter_data_low_counts}
\title{filter minimum total counts and zero count rows}
\usage{
.filter_data_low_counts(data, min_total_counts = 15)
}
\arguments{
\item{data}{data}

\item{min_total_counts}{minimum total counts for variant}
}
\description{
filter minimum total counts and zero count rows
}
