% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace.R
\name{.get_phi_summary}
\alias{.get_phi_summary}
\title{get phi parameter summary (intended for internal use)}
\usage{
.get_phi_summary(fit)
}
\arguments{
\item{fit}{stan output fit object}
}
\description{
get phi parameter summary (intended for internal use)
}
