% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace.R
\name{.get_score_df}
\alias{.get_score_df}
\title{get variant scores dataframe}
\usage{
.get_score_df(lilace_obj, use_positions)
}
\arguments{
\item{lilace_obj}{fitted Lilace object}

\item{use_positions}{whether Lilace was run with use_positions=T}
}
\description{
get variant scores dataframe
}
