% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace.R
\name{.summarize_posteriors}
\alias{.summarize_posteriors}
\title{summarize posterior samples (intended for internal use)}
\usage{
.summarize_posteriors(
  fit,
  data,
  input,
  control_correction,
  control_label,
  use_positions
)
}
\arguments{
\item{fit}{stan output fit object}

\item{data}{data}

\item{input}{stan input list}

\item{control_correction}{logical for whether to do negative control correction}

\item{control_label}{label for negative control variant types}

\item{use_positions}{logical for whether positions used in running Lilace}
}
\description{
summarize posterior samples (intended for internal use)
}
