% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace-package.R
\docType{package}
\name{lilace-package}
\alias{lilace}
\alias{lilace-package}
\title{lilace: Analyze FACS-based Deep Mutational Scanning Data}
\description{
A Bayesian hierarchical model for estimating variant effect sizes and uncertainty from FACS-based DMS data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jermoef/lilace}
  \item \url{http://pimentellab.com/lilace/}
  \item Report bugs at \url{https://github.com/jermoef/lilace/issues}
}

}
\author{
\strong{Maintainer}: Jerome Freudenberg \email{jfreudenberg@ucla.edu} (-15792)

}
\keyword{internal}
