% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace.R
\name{lilace_fit_model}
\alias{lilace_fit_model}
\title{Run Lilace}
\usage{
lilace_fit_model(
  lilace_obj,
  output_dir,
  control_label = "synonymous",
  control_correction = TRUE,
  use_positions = TRUE,
  pseudocount = TRUE,
  seed = NULL,
  min_total_counts = 15,
  n_parallel_chains = 4
)
}
\arguments{
\item{lilace_obj}{initialized lilace object}

\item{output_dir}{output directory to write scores and sampling logs to}

\item{control_label}{label from $data$type column to use as negative controls to score against}

\item{control_correction}{boolean for whether to use negative control scores as bias correction}

\item{use_positions}{boolean for whether to use position hierarchy to improve estimation}

\item{pseudocount}{boolean for whether to add pseudocount (+1 to all counts) for fitting model}

\item{seed}{random seed for sampling process to get exactly reproducible results. A NULL value indicates no fixed seed.}

\item{min_total_counts}{minimum total counts for a variant--anything less will be filtered out}

\item{n_parallel_chains}{number of chains to run in parallel}
}
\value{
lilace object with $scores and $fitted_data entries
}
\description{
Runs Lilace on given data. Lilace will run on $normalized_data if it exists, otherwise it will use $data.
}
\examples{
\dontrun{
lilace_obj <- lilace_fit_model(lilace_obj, output_dir, control_label="synonymous",
                               control_correction=T, use_positions=T, pseudocount=T)
}
}
