% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{lilace_from_counts}
\alias{lilace_from_counts}
\title{Create Lilace object from counts}
\usage{
lilace_from_counts(
  variant_id,
  mutation_type,
  position,
  replicate,
  counts,
  metadata
)
}
\arguments{
\item{variant_id}{a length V vector of unique variant identifiers such as hgvs nomenclature}

\item{mutation_type}{a vector of mutation types (e.g. synonymous, missense, etc).
The label used to identify negative control label should be included in this.}

\item{position}{a vector of residue positions}

\item{replicate}{a vector of replicate ids}

\item{counts}{a (V x K) matrix of bin counts, where K is the number of bins}

\item{metadata}{a matrix or dataframe of any additional variant information to be incorporated into the final lilace output (e.g. wildtype residue)}
}
\value{
a lilace object, with data stored in $data
}
\description{
Formats required information for Lilace to run into a Lilace object. See intro vignette for example usage.
}
