% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{lilace_from_enrich}
\alias{lilace_from_enrich}
\title{Create Lilace object from enrich processed counts output}
\usage{
lilace_from_enrich(file, pheno = "abundance")
}
\arguments{
\item{file}{enrich counts file (see end of intro vignette for example input format)}

\item{pheno}{name of phenotype ("condition" row in counts table)}
}
\value{
a lilace object, with data stored in $data
}
\description{
Create Lilace object from enrich processed counts output
}
