% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{lilace_from_files}
\alias{lilace_from_files}
\title{Create Lilace object from separate count files for each bin and replicate}
\usage{
lilace_from_files(
  file_list,
  variant_id_col = "hgvs",
  position_col = "position",
  mutation_type_col = "type",
  count_col = "count",
  delim = "\\t"
)
}
\arguments{
\item{file_list}{a list containing a separate list of bin count files for each replicate  (see end of intro vignette for example)}

\item{variant_id_col}{column name for variant id}

\item{position_col}{column name for position information}

\item{mutation_type_col}{column name for mutation type information}

\item{count_col}{column name for counts}

\item{delim}{the field separator character (see ?readr::read.delim)}
}
\value{
a lilace object, with data stored in $data
}
\description{
Create Lilace object from separate count files for each bin and replicate
}
