% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{lilace_score_density}
\alias{lilace_score_density}
\title{Generate score histograms/density plots by mutation type}
\usage{
lilace_score_density(
  data,
  savedir,
  type.col = "type",
  score.col = "mean",
  hist = FALSE,
  nbins = 30,
  c.fill = c("#FF7575", "lightgreen", "#7298BF"),
  alpha = 0.5,
  x.text = 10,
  y.text = 10,
  scale.free = FALSE,
  space.free = FALSE,
  ht = 10,
  wd = 8,
  name = "DensityPlot",
  savepdf = TRUE,
  savesvg = TRUE,
  show = TRUE
)
}
\arguments{
\item{data}{Scores data frame. Expected to have columns containing information about
position, control amino acid, mutated amino acid, mutation type, and score.}

\item{savedir}{Character string specifying the directory to save plots.}

\item{type.col}{Column name in `data` for mutation types. Default is `type`.}

\item{score.col}{Column name in `data` for mutation scores. Default is `mean`.}

\item{hist}{Logical indicating whether to plot count histogram or density.
Default is `FALSE`.}

\item{nbins}{Numeric value specifying the number of bins for the histogram.
Default is `30`.}

\item{c.fill}{Vector indicating the fill color for mutation types.}

\item{alpha}{Numeric value between 0-1 indicating the fill transparency.
Default is `0.5`}

\item{x.text}{Numeric value for x-axis text size. Default is `10`.}

\item{y.text}{Numeric value for x-axis text size. Default is `10`.}

\item{scale.free}{Logical indicating whether to make score range proportional.
Default is `FALSE`.}

\item{space.free}{Logical indicating whether to make panel heights variable.
Default is `FALSE`.}

\item{ht}{Numeric value for the height of the saved plot. Default is `10`.}

\item{wd}{Numeric value for the width of the saved plot. Default is `8`.}

\item{name}{Character string specifying the base name of the saved file.}

\item{savepdf}{Logical indicating whether to also save a PDF version of the plot.
Default is `TRUE`.}

\item{savesvg}{logical indicating whether to save as svg}

\item{show}{Logical indicating whether or not to display the plot in the viewer.
Default is `TRUE`.}
}
\value{
NULL.
}
\description{
generates a density plot for visualizing the distribution of scores across different mutation types (same as Rosace scoreDensity function)
}
\examples{
\dontrun{
lilace_score_density(scores, output_dir, score.col="effect",
                     name="score_histogram", hist=T, scale.free=T)
}
}
