% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace.R
\name{.generate_model_input}
\alias{.generate_model_input}
\title{format model input (intended for internal use)}
\usage{
.generate_model_input(
  data,
  control_label,
  pseudocount = F,
  min_total_counts = 15
)
}
\arguments{
\item{data}{data}

\item{control_label}{control label}

\item{pseudocount}{logical for pseudocount}

\item{min_total_counts}{minimum total counts for variant}
}
\description{
format model input (intended for internal use)
}
