% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lilace-package.R
\docType{package}
\name{lilace-package}
\alias{lilace}
\alias{lilace-package}
\title{lilace: Analyze FACS-based Deep Mutational Scanning Data}
\description{
A Bayesian hierarchical model for estimating variant effect sizes and uncertainty from FACS-based DMS data.
}
\author{
\strong{Maintainer}: Jerome Freudenberg \email{jfreudenberg@ucla.edu} (-15792)

}
\keyword{internal}
